requirements.txt
numpy==1.21.1
scipy==1.7.1
torch==1.9.0
torchvision==0.10.0
scikit-image==0.18.3
scikit-learn==0.24.2
matplotlib==3.4.3
nibabel==3.2.1
Pillow==8.3.1
opencv-python==4.5.3
scikit-image==0.18.3
torch==1.9.0
torchvision==0.10.0
torchsummary==1.5.1
timm==0.5.1
